Restructure locales list and fix CI
